Remove hidden mod for .invsee
