+++ Wurst Client/src/tk/wurst_client/command/commands/Invsee.java
@@ -7,12 +7,18 @@
  */
 package tk.wurst_client.command.commands;
 
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.entity.EntityOtherPlayerMP;
+import net.minecraft.client.gui.inventory.GuiInventory;
 import tk.wurst_client.Client;
 import tk.wurst_client.command.Command;
-import tk.wurst_client.module.modules.InvseeCmd;
+import tk.wurst_client.event.EventManager;
+import tk.wurst_client.event.listeners.RenderListener;
 
-public class Invsee extends Command
+public class Invsee extends Command implements RenderListener
 {
+	private String playerName;
+
 	public Invsee()
 	{
 		super("invsee",
@@ -28,8 +34,30 @@ public class Invsee extends Command
 			commandError();
 			return;
 		}
-		InvseeCmd.playerName = args[0];
-		Client.wurst.moduleManager.getModuleFromClass(InvseeCmd.class)
-			.setToggled(true);
+		playerName = args[0];
+		EventManager.addRenderListener(this);
+	}
+	
+	@Override
+	public void onRender()
+	{
+		boolean found = false;
+		for(Object entity : Minecraft.getMinecraft().theWorld.loadedEntityList)
+			if(entity instanceof EntityOtherPlayerMP)
+			{
+				EntityOtherPlayerMP player = (EntityOtherPlayerMP)entity;
+				if(player.getName().equals(playerName))
+				{
+					Client.wurst.chat.message("Showing inventory of "
+						+ player.getName() + ".");
+					Minecraft.getMinecraft().displayGuiScreen(
+						new GuiInventory(player));
+					found = true;
+				}
+			}
+		if(!found)
+			Client.wurst.chat.error("Player not found.");
+		playerName = null;
+		EventManager.removeRenderListener(this);
 	}
 }
