+++ Wurst Client/src/tk/wurst_client/module/ModuleManager.java
@@ -83,7 +83,6 @@ public class ModuleManager
 		activeModules.add(new Home());
 		activeModules.add(new InstantBunker());
 		activeModules.add(new Invisibility());
-		activeModules.add(new InvseeCmd());
 		activeModules.add(new ItemESP());
 		activeModules.add(new Jesus());
 		activeModules.add(new Jetpack());
